# swagger.json
{
  "consumes": [
    "application/json",
    "application/xml"
  ],
  "produces": [
    "application/json",
    "application/xml"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "title": "BrightBird API.",
    "contact": {
      "email": "venus@ipfsconmmunity.com"
    },
    "license": {
      "name": "MIT",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "0.0.1"
  },
  "host": "192.168.200.103:12356",
  "basePath": "/api/v1",
  "paths": {
...
    "/count-plugins": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "plugin"
        ],
        "summary": "Counts the number of plugins.",
        "operationId": "countPluginsReq",
        "responses": {
          "200": {
            "description": "  //todo fix correctstruct",
            "schema": {
              "$ref": "#/definitions/myString"
            }
          },
          "500": {
            "description": "apiError",
            "schema": {
              "$ref": "#/definitions/apiError"
            }
          }
        }
      }
    },
...
    "/failed-tasks": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "failed-tasks"
        ],
        "summary": "Lists the failed tasks.",
        "operationId": "listFailedTasksReq",
        "responses": {
          "200": {
            "description": "  //todo fix correctstruct",
            "schema": {
              "$ref": "#/definitions/myString"
            }
          },
          "500": {
            "description": "apiError",
            "schema": {
              "$ref": "#/definitions/apiError"
            }
          }
        }
      }
    },
...
    "/pass-rate-trends": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "task"
        ],
        "summary": "Gets the pass rate trends for the last 30 days.",
        "operationId": "passRateTrends",
        "responses": {
          "200": {
            "description": "  //todo fix correctstruct",
            "schema": {
              "$ref": "#/definitions/myString"
            }
          },
          "500": {
            "description": "apiError",
            "schema": {
              "$ref": "#/definitions/apiError"
            }
          }
        }
      }
    },
...
    "/success-quantity-trends": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "success-quantity-trends"
        ],
        "summary": "Retrieves the success quantity trends.",
        "operationId": "successQuantityTrendsReq",
        "responses": {
          "200": {
            "description": "  //todo fix correctstruct",
            "schema": {
              "$ref": "#/definitions/myString"
            }
          },
          "500": {
            "description": "apiError",
            "schema": {
              "$ref": "#/definitions/apiError"
            }
          }
        }
      }
    },
...
    "/task-count": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "tasks"
        ],
        "summary": "Retrieves the Statistics of tasks.",
        "operationId": "getTaskCount",
        "responses": {
          "200": {
            "description": "  //todo fix correctstruct",
            "schema": {
              "$ref": "#/definitions/myString"
            }
          },
          "500": {
            "description": "apiError",
            "schema": {
              "$ref": "#/definitions/apiError"
            }
          }
        }
      }
    },
...
    "/test-data": {
      "get": {
        "description": "Lists the amount of tasks for a job in the last 2 weeks.",
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http"
        ],
        "tags": [
          "test-data"
        ],
        "summary": "Lists test data.",
        "operationId": "listTestData",
        "responses": {
          "200": {
            "description": "  //todo fix correctstruct",
            "schema": {
              "$ref": "#/definitions/myString"
            }
          },
          "500": {
            "description": "apiError",
            "schema": {
              "$ref": "#/definitions/apiError"
            }
          }
        }
      }
    },
...
    "/today-pass-rate-ranking": {
      "get": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "task"
        ],
        "summary": "Retrieves the top 3 job pass rates for today.",
        "operationId": "todayPassRateRankingReq",
        "responses": {
          "200": {
            "description": "  //todo fix correctstruct",
            "schema": {
              "$ref": "#/definitions/myString"
            }
          },
          "500": {
            "description": "apiError",
            "schema": {
              "$ref": "#/definitions/apiError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "apiError": {
      "description": "APIError",
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message"
        }
      },
      "x-go-name": "APIError",
      "x-go-package": "github.com/ipfs-force-community/brightbird/models"
    },
...
    "myString": {
      "description": "MyString repretation string",
      "type": "string",
      "x-go-name": "MyString",
      "x-go-package": "github.com/ipfs-force-community/brightbird/models"
    },
...
  },
...
}